Refactor: + remove main submodules to top and remove main module + move unit test to files
