+++ shop/application/src/configuration/application_configuration.rs
@@ -3,7 +3,7 @@ use std::sync::{Arc, Mutex};
 use common::types::main::base::generic_types::AM;
 use lazy_static::lazy_static;
 
-use crate::main::event::kafka_event_publisher_impl::KafkaEventPublisherImpl;
+use crate::event::kafka_event_publisher_impl::KafkaEventPublisherImpl;
 
 // lazy_static! {
 //     pub static ref EVENT_PUBLISHER: AM<EventPublisherImpl<DomainEventEnum>> =
@@ -11,6 +11,6 @@ use crate::main::event::kafka_event_publisher_impl::KafkaEventPublisherImpl;
 // }
 
 lazy_static! {
-    pub static ref EVENT_PUBLISHER: AM<KafkaEventPublisherImpl> =
+    pub(super) static ref EVENT_PUBLISHER: AM<KafkaEventPublisherImpl> =
         Arc::new(Mutex::new(KafkaEventPublisherImpl::default()));
 }
