+++ shop/application/src/configuration/mod.rs
