+++ shop/application/src/configuration/persistence_configuration.rs
@@ -12,29 +12,29 @@ use postgres_persistence::main::{
     postgres_meal_repository::PostgresMealRepository as MealRepository,
 };
 
-use crate::main::configuration::application_configuration::EVENT_PUBLISHER;
+use crate::configuration::application_configuration::EVENT_PUBLISHER;
 
 pub type ORepository = OrderRepository;
 
 lazy_static! {
-    pub static ref MEAL_ID_GENERATOR: AM<MealIdGenerator> = meal_id_generator();
-    pub static ref MEAL_REPOSITORY: AM<MealRepository> = meal_repository();
-    pub static ref ORDER_ID_GENERATOR: AM<OrderIdGenerator> = order_id_generator();
-    pub static ref ORDER_REPOSITORY: AM<OrderRepository> = order_repository();
+    pub(super) static ref MEAL_ID_GENERATOR: AM<MealIdGenerator> = meal_id_generator();
+    pub(super) static ref MEAL_REPOSITORY: AM<MealRepository> = meal_repository();
+    pub(super) static ref ORDER_ID_GENERATOR: AM<OrderIdGenerator> = order_id_generator();
+    pub(super) static ref ORDER_REPOSITORY: AM<OrderRepository> = order_repository();
 }
 
-pub fn meal_id_generator() -> AM<MealIdGenerator> {
+fn meal_id_generator() -> AM<MealIdGenerator> {
     Arc::new(Mutex::new(MealIdGenerator::new(establish_connection())))
 }
 
-pub fn meal_repository() -> AM<MealRepository> {
+fn meal_repository() -> AM<MealRepository> {
     Arc::new(Mutex::new(MealRepository::new(
         establish_connection(),
         EVENT_PUBLISHER.clone(),
     )))
 }
 
-pub fn order_id_generator() -> AM<OrderIdGenerator> {
+fn order_id_generator() -> AM<OrderIdGenerator> {
     Arc::new(Mutex::new(OrderIdGenerator::new()))
 }
