+++ shop/application/src/configuration/swagger_configuration.rs
