+++ shop/application/src/configuration/use_case_configuration.rs
@@ -18,32 +18,32 @@ use usecase::main::{
     },
 };
 
-use crate::main::configuration::persistence_configuration::{
+use crate::configuration::persistence_configuration::{
     ORepository, MEAL_ID_GENERATOR, MEAL_REPOSITORY, ORDER_REPOSITORY,
 };
 
 const GET_ORDERS_MAX_SIZE: usize = 10;
 
 lazy_static! {
-    pub static ref ADD_MEAL_TO_MENU_USE_CASE: Data<AM<AddMealToMenuUseCase>> =
+    pub(super) static ref ADD_MEAL_TO_MENU_USE_CASE: Data<AM<AddMealToMenuUseCase>> =
         Data::new(add_meal_to_menu_use_case()).clone();
-    pub static ref GET_MEAL_BY_ID_USE_CASE: Data<AM<GetMealByIdUseCase>> =
+    pub(super) static ref GET_MEAL_BY_ID_USE_CASE: Data<AM<GetMealByIdUseCase>> =
         Data::new(get_meal_by_id_use_case().clone());
-    pub static ref GET_MENU_USE_CASE: Data<AM<GetMenuUseCase>> =
+    pub(super) static ref GET_MENU_USE_CASE: Data<AM<GetMenuUseCase>> =
         Data::new(get_menu_use_case()).clone();
-    pub static ref REMOVE_MEAL_FROM_MENU_USECASE: Data<AM<RemoveMealFromMenuUseCase>> =
+    pub(super) static ref REMOVE_MEAL_FROM_MENU_USECASE: Data<AM<RemoveMealFromMenuUseCase>> =
         Data::new(remove_meal_from_menu_usecase()).clone();
-    pub static ref CANCEL_ORDER_USECASE: Data<AM<CancelOrderUseCase<ORepository, ORepository>>> =
+    pub(super) static ref CANCEL_ORDER_USECASE: Data<AM<CancelOrderUseCase<ORepository, ORepository>>> =
         Data::new(cancel_order_usecase().clone());
-    pub static ref CONFIRM_ORDER_USECASE: Data<AM<ConfirmOrderUseCase<ORepository, ORepository>>> =
+    pub(super) static ref CONFIRM_ORDER_USECASE: Data<AM<ConfirmOrderUseCase<ORepository, ORepository>>> =
         Data::new(confirm_order_usecase().clone());
-    pub static ref GET_ORDER_BY_ID: Data<AM<GetOrderByIdUseCase<ORepository>>> =
+    pub(super) static ref GET_ORDER_BY_ID: Data<AM<GetOrderByIdUseCase<ORepository>>> =
         Data::new(get_order_by_id_usecase().clone());
-    pub static ref GET_ORDERS_USECASE: Data<AM<GetOrdersUseCase<ORepository>>> =
+    pub(super) static ref GET_ORDERS_USECASE: Data<AM<GetOrdersUseCase<ORepository>>> =
         Data::new(get_orders_usecase().clone());
 }
 
-pub fn add_meal_to_menu_use_case() -> AM<AddMealToMenuUseCase> {
+fn add_meal_to_menu_use_case() -> AM<AddMealToMenuUseCase> {
     let rule = MealAlreadyExistsUsesMealExtractor::new(MEAL_REPOSITORY.clone());
 
     let usecase = AddMealToMenuUseCase::new(
@@ -54,36 +54,36 @@ pub fn add_meal_to_menu_use_case() -> AM<AddMealToMenuUseCase> {
     Arc::new(Mutex::new(usecase))
 }
 
-pub fn get_meal_by_id_use_case() -> AM<GetMealByIdUseCase> {
+fn get_meal_by_id_use_case() -> AM<GetMealByIdUseCase> {
     let usecase = GetMealByIdUseCase::new(MEAL_REPOSITORY.clone());
     Arc::new(Mutex::new(usecase))
 }
 
-pub fn get_menu_use_case() -> AM<GetMenuUseCase> {
+fn get_menu_use_case() -> AM<GetMenuUseCase> {
     let usecase = GetMenuUseCase::new(MEAL_REPOSITORY.clone());
     Arc::new(Mutex::new(usecase))
 }
 
-pub fn remove_meal_from_menu_usecase() -> AM<RemoveMealFromMenuUseCase> {
+fn remove_meal_from_menu_usecase() -> AM<RemoveMealFromMenuUseCase> {
     let usecase = RemoveMealFromMenuUseCase::new(MEAL_REPOSITORY.clone(), MEAL_REPOSITORY.clone());
     Arc::new(Mutex::new(usecase))
 }
 
-pub fn cancel_order_usecase() -> AM<CancelOrderUseCase<ORepository, ORepository>> {
+fn cancel_order_usecase() -> AM<CancelOrderUseCase<ORepository, ORepository>> {
     let usecase = CancelOrderUseCase::new(ORDER_REPOSITORY.clone(), ORDER_REPOSITORY.clone());
     Arc::new(Mutex::new(usecase))
 }
-pub fn confirm_order_usecase() -> AM<ConfirmOrderUseCase<ORepository, ORepository>> {
+fn confirm_order_usecase() -> AM<ConfirmOrderUseCase<ORepository, ORepository>> {
     let usecase = ConfirmOrderUseCase::new(ORDER_REPOSITORY.clone(), ORDER_REPOSITORY.clone());
     Arc::new(Mutex::new(usecase))
 }
 
-pub fn get_order_by_id_usecase() -> AM<GetOrderByIdUseCase<ORepository>> {
+fn get_order_by_id_usecase() -> AM<GetOrderByIdUseCase<ORepository>> {
     let usecase = GetOrderByIdUseCase::new(ORDER_REPOSITORY.clone());
     Arc::new(Mutex::new(usecase))
 }
 
-pub fn get_orders_usecase() -> AM<GetOrdersUseCase<ORepository>> {
+fn get_orders_usecase() -> AM<GetOrdersUseCase<ORepository>> {
     let usecase = GetOrdersUseCase::new(ORDER_REPOSITORY.clone(), || GET_ORDERS_MAX_SIZE + 1);
     Arc::new(Mutex::new(usecase))
 }
