+++ shop/application/src/event/integration_message_publisher.rs
@@ -2,7 +2,7 @@ use async_trait::async_trait;
 use serde::Serialize;
 
 #[async_trait]
-pub trait IntegrationMessagePublisher {
+pub(crate) trait IntegrationMessagePublisher {
     async fn send(
         &self,
         message: impl Serialize + Send + Sync,
