+++ shop/application/src/event/kafka_event_listener_impl.rs
