--- shop/application/src/main/event/mod.rs
@@ -1,5 +1,0 @@
-pub mod event_publisher_impl;
-pub mod integration_message_publisher;
-pub mod kafka_event_listener_impl;
-pub mod kafka_event_publisher_impl;
-pub mod rabbit_message_publisher;
