--- shop/application/src/main/mod.rs
@@ -1,5 +1,0 @@
-pub mod application_startup;
-pub mod configuration;
-pub mod controllers;
-pub mod event;
-pub mod listeners;
