--- shop/application/src/test/event/mod.rs
@@ -1,6 +1,0 @@
-#[cfg(test)]
-pub mod event_publisher_impl_test;
-#[cfg(test)]
-pub mod kafka_event_publisher_impl_test;
-#[cfg(test)]
-pub mod rabbit_message_publisher_test;
